requirements.txt
flash
flask
python-dotenv
openai
supabase
